% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcluster.R
\name{barcluster}
\alias{barcluster}
\title{Plot a Sankey Diagram from a table of unique IDs, and two or more columns indicating membership groups to progress through, in order.}
\usage{
barcluster(
  irl,
  bt,
  alpha = c(0, 0.5, 1),
  beta = 0.1,
  res = 1,
  method = "fast",
  ...
)
}
\arguments{
\item{irl}{Matrix. Principal components matrix, output from \code{BarCluster::irlba_wrap}.}

\item{bt}{Data table. Barcode table of two columns cell IDs ("rn") and barcodes ("Barcode").}

\item{alpha}{Numeric. Alpha parameter or list of values to iterate over. Ranges from 0 to 1. A value of zero reflects only transcriptome edges, a value of one reflects only barcodes.}

\item{beta}{Numeric. Exponent on alpha to adjust curve sloping. Default is 0.1.}

\item{res}{Numeric. Resolution parameter passed to community detection algorithm.}

\item{method}{c("fast", "index"), how the barcode matrix is made, default is fast, if its going too slow try "index".}

\item{...}{Additional arguments passed to \code{RunModularityClustering}.}
}
\value{
A long format data table with three columns, cell ID ("rn"), alpha value, beta value, and cluster assignment ("Group").
}
\description{
Plot a Sankey Diagram from a table of unique IDs, and two or more columns indicating membership groups to progress through, in order.
}
