% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcluster_model.R
\name{barcluster_model}
\alias{barcluster_model}
\title{Return the barcluster network graph for clustering.}
\usage{
barcluster_model(alpha, beta = 0.1, m, nm)
}
\arguments{
\item{alpha}{Numeric. Between zero and one, with zero representing complete transcriptome and 1 representing complete barcode graphs.}

\item{beta}{Numeric. Exponent on alpha to adjust curve sloping. Default is 0.1.}

\item{m}{Object of class \code{dgCMatrix}, the graph of transcriptome edges derived from SNN jaccard index. Usually from \code{Seurat::FindNeighbors}.}

\item{nm}{Output of \code{build_barcode_matrix} or \code{build_barcode_matrix_fast}.}
}
\value{
A sparse matrix of class \code{dgCMatrix} suitable for \code{Seurat:::RunModularityClustering}.
}
\description{
Return the barcluster network graph for clustering.
}
