% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_barcode_matrix.R
\name{build_barcode_matrix}
\alias{build_barcode_matrix}
\title{Build a sparse matrix representing the barcode graph.}
\usage{
build_barcode_matrix(bt, value = NULL)
}
\arguments{
\item{bt}{Data.table. Columns are "rn" (cell ids) and "Barcode".}

\item{value}{Numeric. A value to be used for edge weights within barcodes. Default NULL will use the reciprocal of barcode size.}
}
\value{
A sparse matrix of class \code{dgCMatrix} suitable for \code{clonocluster_model} or \code{Seurat::FindClusters}.
}
\description{
Build a sparse matrix representing the barcode graph.
}
