% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vectors.R
\docType{data}
\name{c25}
\alias{c25}
\title{internal function, vector of colors for discrete plots.}
\format{
An object of class \code{character} of length 25.
}
\usage{
c25
}
\description{
internal function, vector of colors for discrete plots.
}
\keyword{datasets}
