% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_confusion.R
\name{cast_confusion}
\alias{cast_confusion}
\title{Generate confusion matrix for two clusterings, then return summary statistics.}
\usage{
cast_confusion(clusters, barcodes)
}
\arguments{
\item{clusters}{Data.table. Columns are cell ids and cluster assignment.}

\item{barcodes}{Data.table. Columns are cell ids and barcode assignment.}
}
\value{
A data.table with summary statistics from the contingency tables, for each barcode, the cluster(s) with the highest number of true positives is returned.
}
\description{
Generate confusion matrix for two clusterings, then return summary statistics.
}
