% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_vectors.R
\docType{data}
\name{cw_colors}
\alias{cw_colors}
\title{Initialize vector of plot colors published by Claus Wilke.}
\format{
An object of class \code{character} of length 14.
}
\usage{
cw_colors
}
\description{
Initialize vector of plot colors published by Claus Wilke.
}
\keyword{datasets}
