% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dt2m}
\alias{dt2m}
\title{Convert a data.table to matrix.}
\usage{
dt2m(dt)
}
\arguments{
\item{dt}{A data.table to be converted to a matrix.}
}
\value{
A matrix from the data.table, with the first table column as rownames of the matrix.
}
\description{
Convert a data.table to matrix.
}
