% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_warp.R
\name{engage_warp}
\alias{engage_warp}
\title{Warp principal components and return umap.}
\usage{
engage_warp(pca, bt, s, ...)
}
\arguments{
\item{pca}{Matrix. Principal components matrix, output from \code{BarCluster::irlba_wrap}.}

\item{bt}{Data table. Barcode table of two columns cell IDs ("rn") and barcodes ("Barcode").}

\item{s}{Numeric. Warp factor, from 0 to 10.}

\item{...}{Other arguments passed to \code{umap_matrix}.}
}
\value{
UMAP output from the warped PCA.
}
\description{
Warp principal components and return umap.
}
