% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCR_wrap.R
\name{Find_Markers_ROC}
\alias{Find_Markers_ROC}
\title{Fast AUC calculation wrapper for clusters.}
\usage{
Find_Markers_ROC(dl, cm, n_threads = 1)
}
\arguments{
\item{dl}{A data.table of cluster assignments, output from \code{clonocluster}, with minimum columns c("rn", "Group", "alpha").}

\item{cm}{Matrix. A count matrix.}

\item{n_threads}{Integer. Passed to \code{mc.cores} for \code{parallel::mclapply}}
}
\value{
A table of marker AUCs and thresholds for all clusters.
}
\description{
Fast AUC calculation wrapper for clusters.
}
