% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{FindAllMarkers_Seurat}
\alias{FindAllMarkers_Seurat}
\title{A wrapper to call Seurat to FindAllMarkers on given groups.}
\usage{
FindAllMarkers_Seurat(so, groupt, ...)
}
\arguments{
\item{so}{A Seurat object or path to one.}

\item{...}{All arguments passed to \code{Seurat::FindAllMarkers}.}

\item{clust}{A table of cellIDs and group assignments.}
}
\value{
A data.table of the output.
}
\description{
A wrapper to call Seurat to FindAllMarkers on given groups.
}
