% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irlba_wrap.R
\name{irlba_wrap}
\alias{irlba_wrap}
\title{A wrapper to run approximate PCA on a count matrix.}
\usage{
irlba_wrap(gt, npc = 100, seed_use = 42)
}
\arguments{
\item{gt}{Matrix. Rows are cell IDs, columns are gene counts.}

\item{npc}{Integer. Number of PCs to compute with \code{irlba::irlba}, default is 100.}

\item{seed_use}{Integer. RNG seed for PCA, default is 42.}
}
\value{
A matrix with cell as rows, PCs as features/columns.
}
\description{
A wrapper to run approximate PCA on a count matrix.
}
