% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_alluvia.R
\name{Plot_alluvia}
\alias{Plot_alluvia}
\title{Plot a Sankey Diagram from a table of unique IDs, and two or more columns indicating membership groups to progress through, in order.}
\usage{
Plot_alluvia(
  dl,
  bl,
  title = "",
  xlab = "",
  ylab = "",
  flow_alpha = 0.2,
  node_alpha = 1,
  stack_colors = TRUE,
  border_size = 2,
  label_nodes = TRUE,
  label_size = 2,
  ltype = "label",
  reverse = TRUE,
  cols = ClonoCluster::c25
)
}
\arguments{
\item{dl}{A data table, with three columns, c("rn", "Group", "alpha").}

\item{bl}{A data table, with two columns, c("rn", "Barcode")}

\item{title}{Character. Passed to \code{ggtitle}.}

\item{xlab}{Character. Passed to \code{xlab}.}

\item{ylab}{Character. Passed to \code{ylab}.}

\item{flow_alpha}{Numeric. Alpha value for the alluvial ribbons.}

\item{node_alpha}{Numeric. Alpha value for the strata, passed to geom_rect.}

\item{stack_colors}{Boolean. Should nodes be represented as stacked colors representing the contributing alluvia? Default TRUE.}

\item{border_size}{Numeric. Rectangle thickness around the sankey nodes. Default 2.}

\item{label_nodes}{Boolean. Should clusters be labeled with geom_label? Default TRUE.}

\item{label_size}{Numeric. size of geom_text labels font.}

\item{ltype}{One of c("label", "text"). Passed as the geom to annotation. Default is "label"}

\item{reverse}{Boolean. Default TRUE, return plot colored by clusters and barcodes.}

\item{cols}{Character. Vector of colors values, default is c25.}
}
\value{
Returns a ggplot object (or two if reverse == TRUE) of the Sankey Plot.
}
\description{
Plot a Sankey Diagram from a table of unique IDs, and two or more columns indicating membership groups to progress through, in order.
}
