% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_alluvia_counts.R
\name{Plot_alluvia_counts}
\alias{Plot_alluvia_counts}
\title{Plot a Sankey Diagram from a table of unique IDs, and two or more columns indicating membership groups to progress through, in order.}
\usage{
Plot_alluvia_counts(
  dl,
  counts,
  title = "",
  xlab = "",
  ylab = "",
  flow_alpha = 0.2,
  node_alpha = 1,
  border_size = 2,
  label_nodes = TRUE,
  label_size = 2,
  ltype = "text",
  col_start = "gray100",
  col_end = "darkblue"
)
}
\arguments{
\item{dl}{A data table, with three columns, c("rn", "Group", "alpha").}

\item{counts}{A matrix with one count column and rownames equivalent to "rn".}

\item{title}{Character. Passed to \code{ggtitle}.}

\item{xlab}{Character. Passed to \code{xlab}.}

\item{ylab}{Character. Passed to \code{ylab}.}

\item{flow_alpha}{Numeric. Alpha value for the alluvial ribbons.}

\item{node_alpha}{Numeric. Alpha value for the strata, passed to geom_rect.}

\item{border_size}{Numeric. Rectangle thickness around the sankey nodes. Default 2.}

\item{label_nodes}{Boolean. Should clusters be labeled with geom_label? Default TRUE.}

\item{label_size}{Numeric. size of geom_text labels font.}

\item{ltype}{One of c("label", "text"). Passed as the geom to annotation. Default is "label"}

\item{col_start}{Character. Starting value for color gradient, default is "gray100".}

\item{col_end}{Character. Ending value for color gradient, default is "darkblue".}
}
\value{
Returns a ggplot object of the Sankey Plot, alluvia are directly proportional to counts and node colors are the average of all contributing cells.
}
\description{
Plot a Sankey Diagram from a table of unique IDs, and two or more columns indicating membership groups to progress through, in order.
}
