% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_alluvia_track.R
\name{Plot_alluvia_track}
\alias{Plot_alluvia_track}
\title{Plot a Sankey Diagram and track a list of cellIDs.}
\usage{
Plot_alluvia_track(
  dl,
  ids,
  title = "",
  xlab = "",
  ylab = "",
  flow_alpha = 0.2,
  node_alpha = 1,
  border_size = 2,
  label_nodes = TRUE,
  label_size = 2,
  ltype = "text",
  cols = BarCluster::cw_colors,
  alluvia_cols = cols,
  col2 = "gray100",
  orientation = "bottom"
)
}
\arguments{
\item{dl}{A data table, with three columns, c("rn", "Group", "alpha"), from \code{barcluster}.}

\item{ids}{List. A list of character vectors containing rownames in dl (cellIDs) for each group to be tracked.}

\item{title}{Character. Passed to \code{ggtitle}.}

\item{xlab}{Character. Passed to \code{xlab}.}

\item{ylab}{Character. Passed to \code{ylab}.}

\item{flow_alpha}{Numeric. Alpha value for the alluvial ribbons.}

\item{node_alpha}{Numeric. Alpha value for the strata, passed to geom_rect.}

\item{border_size}{Numeric. Rectangle thickness around the sankey nodes. Default 2.}

\item{label_nodes}{Boolean. Should clusters be labeled with geom_label? Default TRUE.}

\item{label_size}{Numeric. size of geom_text labels font.}

\item{ltype}{One of c("label", "text"). Passed as the geom to annotation. Default is "label"}

\item{cols}{Character. Vector of colors values of length equal to the number of groups to be tracked, node colors.}

\item{alluvia_cols}{Character. Vector of colors values of length equal to the number of groups to be tracked, alluvium colors, default is same as \code{cols}.}

\item{col2}{Character. Color of untracked cells, default "gray100".}

\item{orientation}{One of c("top", "bottom"), where colored rectangles will be stacked on the nodes. Default is "bottom".}
}
\value{
Returns a ggplot object of the Sankey Plot.
}
\description{
Plot a Sankey Diagram and track a list of cellIDs.
}
