% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCR_wrap.R
\name{ROCR_wrap}
\alias{ROCR_wrap}
\title{Fast AUC calculation.}
\usage{
ROCR_wrap(x, y, return_curve = FALSE)
}
\arguments{
\item{x}{Numeric. First vector of values.}

\item{y}{Numeric. Second vector of values.}

\item{return_curve}{Logical. Default FALSE. Return the ROC curve values.}
}
\value{
AUC value (always > 0.5) or ROC curve values.
}
\description{
Fast AUC calculation.
}
