% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{RunModularityClustering}
\alias{RunModularityClustering}
\title{internal function to run louvain algorithm lifted from \code{Seurat}.}
\usage{
RunModularityClustering(
  SNN = matrix(),
  modularity = 1,
  resolution = 1,
  algorithm = 1,
  n.start = 1,
  n.iter = 10,
  random.seed = 42,
  print.output = TRUE,
  temp.file.location = NULL,
  edge.file.name = NULL
)
}
\description{
internal function to run louvain algorithm lifted from \code{Seurat}.
}
