% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{RunModularityClusteringCpp}
\alias{RunModularityClusteringCpp}
\title{internal function to call C code from\code{Seurat} for community detection.}
\usage{
RunModularityClusteringCpp(
  SNN,
  modularityFunction,
  resolution,
  algorithm,
  nRandomStarts,
  nIterations,
  randomSeed,
  printOutput,
  edgefilename
)
}
\description{
internal function to call C code from\code{Seurat} for community detection.
}
