% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tdt}
\alias{tdt}
\title{Flip a count matrix data.table.}
\usage{
tdt(dt)
}
\arguments{
\item{dt}{A data.table to be flipped.}
}
\value{
A data table that has been transposed
}
\description{
Flip a count matrix data.table.
}
