% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{data}
\name{ttheme}
\alias{ttheme}
\title{internal function, plot theming.}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 3.
}
\usage{
ttheme
}
\description{
internal function, plot theming.
}
\keyword{datasets}
