% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_matrix.R
\name{umap_matrix}
\alias{umap_matrix}
\title{Run \code{uwot::umap} on a matrix using Seurat v2.0 defaults.}
\usage{
umap_matrix(
  ce,
  metric = "cosine",
  n_neighbors = 30L,
  n_components = 2L,
  min_dist = 0.3,
  learning_rate = 1,
  spread = 1,
  set_op_mix_ratio = 1,
  local_connectivity = 1L,
  repulsion_strength = 1,
  negative_sample_rate = 5L,
  init = "spectral",
  seed_use = 42,
  ...
)
}
\arguments{
\item{ce}{Matrix. Input matrix to reduce.}

\item{metric}{Character. Default "cosine", other options c("euclidean", "manhattan", "hamming", "correlation", "categorical").}

\item{n_neighbors}{Integer. Nearest neighbors to search, default 30.}

\item{n_components}{Integer. Number of UMAP dimensions to return.}

\item{min_dist}{Numeric. Minimum distance between points, default 0.3.}

\item{learning_rate}{Numeric. Initial learning rate for optimization. Default 1.}

\item{spread}{Numeric. Effective scale of embedded points. Default 1.}

\item{set_op_mix_ratio}{Numeric. Default is 1, pure fuzzy union.}

\item{local_connectivity}{Integer. Number of nearest neighbors to consider at local level. Default 1.}

\item{repulsion_strength}{Numeric. Weighting applied to negative sampling. Default 1.}

\item{negative_sample_rate}{Integer. Ratio of negative to positive samples to use for optimization of embedding. Default 5.}

\item{init}{Character. Default "spectral". Type of initialization for coordinates, other options c("normlaplacian", "random", "lvrandom", "laplacian", "pca", "spca", "agspectral") or matrix of coordinates.}

\item{seed_use}{Numeric. RNG seed passed to \code{set_seed}, default is 42.}
}
\value{
A data table of the input rownames, UMAP components, and seed used.
}
\description{
Run \code{uwot::umap} on a matrix using Seurat v2.0 defaults.
}
